ref: re-construct event module
